requirements.txt
netifaces
zeroconf
biplist
pycryptodome
hexdump
srptools
hkdf
pynacl
cryptography
requests
av